Refactor pom and folder structure to fit with other scrapers
