+++ czytampl_scraper/src/main/java/com/blackbook/czytamplscraper/scraper/WebReader.java
@@ -1,13 +1,13 @@
-package com.blackbook.czytampl.scraper;
+package com.blackbook.czytamplscraper.scraper;
 
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
 
 import java.io.IOException;
 
-class WebReader {
+public class WebReader {
 
-    Document getDocumentFromWebPage(String url){
+    public Document getDocumentFromWebPage(String url){
         try {
             return Jsoup.connect(url).get();
         } catch (IOException e) {
