+++ czytampl_scraper/src/test/java/com/blackbook/czytamplscraper/.keep
@@ -1,0 +1,1 @@
+To store folder structure
